Added authorization checks to various Controllers.

+++ app/Http/Controllers/PersonController.php
@@ -52,11 +52,15 @@ class PersonController extends Controller
 
     public function create()
     {
+        Gate::authorize('create', Person::class);
+
         return Inertia::render('admin/person/edit/index');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', Person::class);
+
         $dataForm = $request->all();
 
         $person = Person::create($dataForm);
@@ -74,6 +78,8 @@ class PersonController extends Controller
 
     public function edit(Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         $person->load([
             'artworks',
             'reviews',
@@ -86,6 +92,8 @@ class PersonController extends Controller
 
     public function update(Request $request, Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         $dataForm = $request->all();
 
         $person->update($dataForm);
@@ -103,6 +111,8 @@ class PersonController extends Controller
 
     public function editImages(Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         return Inertia::render('admin/person/edit/images', [
             'person' => new PersonResource($person),
         ]);
@@ -110,6 +120,8 @@ class PersonController extends Controller
 
     public function updateImages(Request $request, Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         try {
             $this->handleImageUpdate($request, $person);
 
@@ -126,6 +138,8 @@ class PersonController extends Controller
 
     public function editChronology(Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         return Inertia::render('admin/person/edit/chronology', [
             'person' => new PersonResource($person),
         ]);
@@ -133,6 +147,8 @@ class PersonController extends Controller
 
     public function updateChronology(Request $request, Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         try {
             $this->handleContentUpdate($request, $person);
 
@@ -149,6 +165,8 @@ class PersonController extends Controller
 
     public function editContent(Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         return Inertia::render('admin/person/edit/content', [
             'person' => new PersonResource($person),
         ]);
@@ -156,6 +174,8 @@ class PersonController extends Controller
 
     public function updateContent(Request $request, Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         try {
             $this->handleContentUpdate($request, $person);
 
@@ -172,6 +192,8 @@ class PersonController extends Controller
 
     public function editSources(Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         $person->load('sources');
 
         return Inertia::render('admin/person/edit/sources', [
@@ -181,6 +203,8 @@ class PersonController extends Controller
 
     public function updateSources(Request $request, Person $person)
     {
+        Gate::authorize('update', Person::class);
+
         $this->syncUuids($request->sources_uuids, $person->sources());
 
         session()->flash('success', true);
@@ -192,6 +216,8 @@ class PersonController extends Controller
 
     public function destroy(Person $person)
     {
+        Gate::authorize('delete', Person::class);
+
         $person->delete();
 
         session()->flash('success', true);
@@ -203,6 +229,8 @@ class PersonController extends Controller
 
     public function fetchSelectOptions(Request $request)
     {
+        Gate::authorize('view', Person::class);
+
         return (new SearchController())->fetchMulti(
             $request->merge([
                 'limit' => 5,
